Proposal
Here I didn't go with reinventing the wheel, relied to lib50-c as the backbone.
Whitespace cut-off is carried in get_string method.
I've had in mind the restriction of C++99 all the time.
Want to hear if this design passes the bar. If so, GC and some other features which will improve the robustness can be easily added later on. 

-Ivan